Download Information on Box Office Results for Movies

Download Information on Box Office Results for Movies

boxoffice(dates, site = c("mojo", "numbers"), top_n = NULL)

Arguments

dates
A vector of dates to scrape

site
Whether you want to get data from boxofficemojo.com or the-numbers.com. Accepts inputs of "numbers" (default) or "mojo".

top_n
The number of results to return for each day. If NULL (default) returns all results, otherwise just top n results (e.g. top_n = 5, returns 5 top movies per date).

Value

Data frame returning info on the name of the movie, its daily gross, gross-to-date, and gross-per-theater for each date inputted.

Examples

# Uses the-numbers.com website.
boxoffice(dates = as.Date("2017-12-25"))

#>                             movie          distributor    gross percent_change
#> 1    Star Wars Ep. VIII: The Las…          Walt Disney 27459557             56
#> 2  Jumanji: Welcome to the Jungle        Sony Pictures 19138553            111
#> 3                 Pitch Perfect 3            Universal  6496365            147
#> 4            The Greatest Showman     20th Century Fox  5604273            162
#> 5                            Coco          Walt Disney  2805466            113
#> 6                       Ferdinand     20th Century Fox  2801780             84
#> 7                      Downsizing   Paramount Pictures  2715843            135
#> 8      All the Money in the World        Sony Pictures  2565393             NA
#> 9                  Father Figures         Warner Bros.  2192067            200
#> 10                   Darkest Hour       Focus Features  1622510             73
#> 11             The Shape of Water      Fox Searchlight  1419891             95
#> 12                   Molly’s Game    STX Entertainment  1172036             NA
#> 13                         Wonder            Lionsgate   710214             77
#> 14                 Daddy’s Home 2   Paramount Pictures   544612            149
#> 15                 Justice League         Warner Bros.   435091             76
#> 16                      Lady Bird                  A24   415989             60
#> 17                 Thor: Ragnarok          Walt Disney   329457             64
#> 18            The Disaster Artist                  A24   325057             77
#> 19                       The Star        Sony Pictures   238704            -13
#> 20                       The Post     20th Century Fox   236046             35
#> 21                       I, Tonya                 Neon   210755            109
#> 22   Three Billboards Outside Eb…      Fox Searchlight   207975             42
#> 23   Murder on the Orient Express     20th Century Fox   203411             58
#> 24                 Phantom Thread       Focus Features   127272             NA
#> 25           A Bad Moms Christmas    STX Entertainment    60302            122
#> 26                   Wonder Wheel       Amazon Studios    42778             67
#> 27 The Man Who Invented Christmas      Bleecker Street    33190             48
#> 28                           Jane     Abramorama Films    27215             NA
#> 29                       Geostorm         Warner Bros.    16062            169
#> 30                 Only the Brave        Sony Pictures     9790            145
#> 31                       Hostiles Entertainment Studi…     9606             65
#> 32            The Florida Project                  A24     9430             46
#> 33              Blade Runner 2049         Warner Bros.     8433             52
#> 34          Roman J. Israel, Esq.        Sony Pictures     4786             63
#> 35   Same Kind of Different as Me Pure Flix Entertain…     3069            135
#> 36             Victoria and Abdul       Focus Features     2840            114
#> 37                      Big Sonia       Argot Pictures     2311             53
#> 38                       Marshall            Open Road     1937             83
#> 39                The Breadwinner                GKIDS      792             27
#> 40             Let There Be Light   Atlas Distribution      736              1
#> 41   The Killing of a Sacred Deer                  A24      602              3
#> 42                         Thelma          The Orchard      593             56
#> 43         BPM (Beats per Minute)          The Orchard      160            -35
#> 44   Birdboy: The Forgotten Chil…                GKIDS      117            -25
#>    theaters per_theater total_gross days       date
#> 1      4232        6489   395627411   11 2017-12-25
#> 2      3765        5083    71913848    6 2017-12-25
#> 3      3447        1885    26424890    4 2017-12-25
#> 4      3006        1864    19008847    6 2017-12-25
#> 5      2111        1329   164307743   34 2017-12-25
#> 6      3630         772    29601410   11 2017-12-25
#> 7      2668        1018     7670130    4 2017-12-25
#> 8      2074        1237     2565393    1 2017-12-25
#> 9      2902         755     5479518    4 2017-12-25
#> 10      806        2013     8367532   34 2017-12-25
#> 11      730        1945     8986046   25 2017-12-25
#> 12      271        4325     1172036    1 2017-12-25
#> 13     1130         629   115654826   39 2017-12-25
#> 14     1073         508    99751582   46 2017-12-25
#> 15     1101         395   223128877   39 2017-12-25
#> 16      402        1035    28715314   53 2017-12-25
#> 17      701         470   309454093   53 2017-12-25
#> 18      517         629    16020034   25 2017-12-25
#> 19     1106         216    38594870   39 2017-12-25
#> 20        9       26227      762057    4 2017-12-25
#> 21       42        5018     1277150   18 2017-12-25
#> 22      264         788    22949085   46 2017-12-25
#> 23      361         563    99444878   46 2017-12-25
#> 24        4       31818      127272    1 2017-12-25
#> 25      275         219    71571546   55 2017-12-25
#> 26      113         379     1205919   25 2017-12-25
#> 27      146         227     5419083   34 2017-12-25
#> 28       19        1432     1432348   67 2017-12-25
#> 29      110         146    33473188   67 2017-12-25
#> 30       86         114    18106612   67 2017-12-25
#> 31        3        3202       32455    4 2017-12-25
#> 32       43         219     5204890   81 2017-12-25
#> 33       75         112    91564672   81 2017-12-25
#> 34       49          98    11825352   39 2017-12-25
#> 35       45          68     6370123   67 2017-12-25
#> 36       24         118    22205730   95 2017-12-25
#> 37        3         770       94741   39 2017-12-25
#> 38       58          33     9450507   74 2017-12-25
#> 39       10          79      210278   39 2017-12-25
#> 40        8          92     7205484   60 2017-12-25
#> 41        8          75     2289084   67 2017-12-25
#> 42        9          66      127398   46 2017-12-25
#> 43        1         160       92117   67 2017-12-25
#> 44        1         117       12620   11 2017-12-25

# Uses boxofficemojo.com website.
boxoffice(dates = as.Date("2017-12-25"), site = "mojo")

#> The terms of use for boxofficemojo.com does not permit scraping without their written permission. If you do not have written permission, please ask them for it or change the site parameter to 'numbers' to use the-numbers.com which does not forbid scraping without permission.

#>                                        movie
#> 1    Star Wars: Episode VIII - The Last Jedi
#> 2             Jumanji: Welcome to the Jungle
#> 3                            Pitch Perfect 3
#> 4                       The Greatest Showman
#> 5                                       Coco
#> 6                                  Ferdinand
#> 7                                 Downsizing
#> 8                 All the Money in the World
#> 9                             Father Figures
#> 10                              Darkest Hour
#> 11                        The Shape of Water
#> 12                              Molly's Game
#> 13                           Tiger Zinda Hai
#> 14                                    Wonder
#> 15                            Daddy's Home 2
#> 16                            Justice League
#> 17                                 Lady Bird
#> 18                            Thor: Ragnarok
#> 19                       The Disaster Artist
#> 20                                  The Star
#> 21                                  The Post
#> 22                      Call Me by Your Name
#> 23                                  I, Tonya
#> 24 Three Billboards Outside Ebbing, Missouri
#> 25              Murder on the Orient Express
#> 26                            Phantom Thread
#> 27                                     Youth
#> 28       Along with the Gods: The Two Worlds
#> 29                      A Bad Moms Christmas
#> 30                              Wonder Wheel
#> 31            The Man Who Invented Christmas
#> 32                                  Geostorm
#> 33                            Loving Vincent
#> 34                            Only the Brave
#> 35                                  Hostiles
#> 36                       The Florida Project
#> 37                         Blade Runner 2049
#> 38                                The Square
#> 39                                 Happy End
#> 40              The Thousand Faces of Dunjia
#> 41                     Roman J. Israel, Esq.
#> 42              Same Kind of Different as Me
#> 43                          Victoria & Abdul
#> 44        The Polar Express2017 IMAX Release
#> 45                    The Other Side of Hope
#> 46                                 Big Sonia
#> 47                             The Swindlers
#> 48                             The Lucky Man
#> 49                                  Marshall
#> 50                           The Breadwinner
#> 51                          My Friend Dahmer
#> 52                        Let There Be Light
#> 53              The Killing of a Sacred Deer
#> 54                                    Thelma
#> 55                                 Novitiate
#> 56                           The Paris Opera
#> 57                    BPM (Beats Per Minute)
#> 58                                 Permanent
#> 59           Birdboy: The Forgotten Children
#>                              distributor    gross percent_change theaters
#> 1    Walt Disney Studios Motion Pictures 27459557           55.8     4232
#> 2                Sony Pictures Releasing 19138553          111.5     3765
#> 3                     Universal Pictures  6496365          146.7     3447
#> 4                                    Fox  5604273          161.8     3006
#> 5    Walt Disney Studios Motion Pictures  2805466          112.6     2111
#> 6                                    Fox  2801780           83.7     3630
#> 7                     Paramount Pictures  2715843          135.0     2668
#> 8                          Concorde Film  2565393             NA     2074
#> 9                           Warner Bros.  2192067          199.7     2902
#> 10                        Focus Features  1622510           73.0      806
#> 11              Fox Searchlight Pictures  1419891           95.4      730
#> 12                     STX Entertainment  1172036             NA      271
#> 13                                     -   785745           33.3      299
#> 14                             Lionsgate   710214           76.6      794
#> 15                    Paramount Pictures   544612          149.5     1073
#> 16                          Warner Bros.   435091           75.9     1101
#> 17                                   A24   415989           60.4      372
#> 18   Walt Disney Studios Motion Pictures   329457           64.3      701
#> 19                                   A24   325057           77.2      488
#> 20               Sony Pictures Releasing   238704          -12.7     1106
#> 21                                   Fox   236046           35.2        9
#> 22                Sony Pictures Classics   229888             NA      114
#> 23                                  Neon   220806          108.3       43
#> 24              Fox Searchlight Pictures   207975           42.5      264
#> 25                                   Fox   203411           58.3      361
#> 26                        Focus Features   127272             NA        4
#> 27          China Lion Film Distribution   116922             NA       34
#> 28             Well Go USA Entertainment    99644             NA       17
#> 29                     STX Entertainment    60302          121.9      275
#> 30                        Amazon Studios    42778           67.1      113
#> 31                 Bleecker Street Media    33190           47.7      146
#> 32                          Warner Bros.    16062          169.5      110
#> 33               Good Deed Entertainment    12235             NA       55
#> 34               Sony Pictures Releasing     9790          145.2       86
#> 35 Entertainment Studios Motion Pictures     9606           64.5        3
#> 36                                   A24     9430           45.9       43
#> 37                          Warner Bros.     8433           51.8       75
#> 38                     Magnolia Pictures     7250             NA       17
#> 39                Sony Pictures Classics     7131             NA        3
#> 40             Well Go USA Entertainment     7092             NA        4
#> 41               Sony Pictures Releasing     4786           62.7       49
#> 42               Pure Flix Entertainment     3069          135.2       45
#> 43                        Focus Features     2840          114.3       24
#> 44                          Warner Bros.     2603          -80.7       NA
#> 45                           Janus Films     2386             NA        4
#> 46                        Argot Pictures     2311           52.5        3
#> 47             Well Go USA Entertainment     2162             NA        1
#> 48                      Indican Pictures     1948             NA        3
#> 49                  Open Road Films (II)     1937           82.9       58
#> 50                                 GKIDS      792           26.5       10
#> 51                              FilmRise      784             NA        3
#> 52            Atlas Distribution Company      736            1.1        8
#> 53                                   A24      602            3.4        8
#> 54                           The Orchard      593           56.1        9
#> 55                Sony Pictures Classics      483             NA        8
#> 56                         Film Movement      177             NA        2
#> 57                           The Orchard      160          -35.5        1
#> 58                     Magnolia Pictures      141             NA        3
#> 59                                 GKIDS      117          -25.0        1
#>    per_theater total_gross days       date
#> 1         6488   395627411   11 2017-12-25
#> 2         5083    71913848    6 2017-12-25
#> 3         1884    26424890    4 2017-12-25
#> 4         1864    19008847    6 2017-12-25
#> 5         1328   164307743   34 2017-12-25
#> 6          771    29601410   11 2017-12-25
#> 7         1017     7670130    4 2017-12-25
#> 8         1236     2565393    1 2017-12-25
#> 9          755     5479518    4 2017-12-25
#> 10        2013     8367532   34 2017-12-25
#> 11        1945     8986046   25 2017-12-25
#> 12        4324     1172036    1 2017-12-25
#> 13        2627     2572853    4 2017-12-25
#> 14         894   115654826   39 2017-12-25
#> 15         507    99751582   46 2017-12-25
#> 16         395   223128877   39 2017-12-25
#> 17        1118    28715314   53 2017-12-25
#> 18         469   309454093   53 2017-12-25
#> 19         666    16020034   25 2017-12-25
#> 20         215    38594870   39 2017-12-25
#> 21       26227      762057    4 2017-12-25
#> 22        2016     3262454   32 2017-12-25
#> 23        5135     1306911   18 2017-12-25
#> 24         787    22949085   46 2017-12-25
#> 25         563    99444878   46 2017-12-25
#> 26       31818      127272    1 2017-12-25
#> 27        3438     1157273   11 2017-12-25
#> 28        5861      285190    4 2017-12-25
#> 29         219    71571546   55 2017-12-25
#> 30         378     1205919   25 2017-12-25
#> 31         227     5419083   34 2017-12-25
#> 32         146    33473188   67 2017-12-25
#> 33         222     6110430   95 2017-12-25
#> 34         113    18106612   67 2017-12-25
#> 35        3202       32455    4 2017-12-25
#> 36         219     5204890   81 2017-12-25
#> 37         112    91564672   81 2017-12-25
#> 38         426     1176775   60 2017-12-25
#> 39        2377       30222    4 2017-12-25
#> 40        1773      109690   11 2017-12-25
#> 41          97    11825352   39 2017-12-25
#> 42          68     6370123   67 2017-12-25
#> 43         118    22205730   95 2017-12-25
#> 44          NA      843610   39 2017-12-25
#> 45         596      118777   25 2017-12-25
#> 46         770       94741   39 2017-12-25
#> 47        2162      240979   25 2017-12-25
#> 48         649        6904    5 2017-12-25
#> 49          33     9450507   74 2017-12-25
#> 50          79      210278   39 2017-12-25
#> 51         261     1293780   53 2017-12-25
#> 52          92     7203504   60 2017-12-25
#> 53          75     2289085   67 2017-12-25
#> 54          65      128627   46 2017-12-25
#> 55          60      563231   60 2017-12-25
#> 56          88       51772   69 2017-12-25
#> 57         160       92117   67 2017-12-25
#> 58          47        9238   11 2017-12-25
#> 59         117       12620   11 2017-12-25

# Returns only top 10 (daily) grossing movies
boxoffice(dates = as.Date("2017-12-25"), top_n = 10)

#>                             movie        distributor    gross percent_change
#> 1    Star Wars Ep. VIII: The Las…        Walt Disney 27459557             56
#> 2  Jumanji: Welcome to the Jungle      Sony Pictures 19138553            111
#> 3                 Pitch Perfect 3          Universal  6496365            147
#> 4            The Greatest Showman   20th Century Fox  5604273            162
#> 5                            Coco        Walt Disney  2805466            113
#> 6                       Ferdinand   20th Century Fox  2801780             84
#> 7                      Downsizing Paramount Pictures  2715843            135
#> 8      All the Money in the World      Sony Pictures  2565393             NA
#> 9                  Father Figures       Warner Bros.  2192067            200
#> 10                   Darkest Hour     Focus Features  1622510             73
#>    theaters per_theater total_gross days       date
#> 1      4232        6489   395627411   11 2017-12-25
#> 2      3765        5083    71913848    6 2017-12-25
#> 3      3447        1885    26424890    4 2017-12-25
#> 4      3006        1864    19008847    6 2017-12-25
#> 5      2111        1329   164307743   34 2017-12-25
#> 6      3630         772    29601410   11 2017-12-25
#> 7      2668        1018     7670130    4 2017-12-25
#> 8      2074        1237     2565393    1 2017-12-25
#> 9      2902         755     5479518    4 2017-12-25
#> 10      806        2013     8367532   34 2017-12-25

# Uses the dates of Christmas and New Years Eve 2017
boxoffice(dates = as.Date(c("2017-12-25", "2017-12-31")))

#>                             movie          distributor    gross percent_change
#> 1    Star Wars Ep. VIII: The Las…          Walt Disney 27459557             56
#> 2  Jumanji: Welcome to the Jungle        Sony Pictures 19138553            111
#> 3                 Pitch Perfect 3            Universal  6496365            147
#> 4            The Greatest Showman     20th Century Fox  5604273            162
#> 5                            Coco          Walt Disney  2805466            113
#> 6                       Ferdinand     20th Century Fox  2801780             84
#> 7                      Downsizing   Paramount Pictures  2715843            135
#> 8      All the Money in the World        Sony Pictures  2565393             NA
#> 9                  Father Figures         Warner Bros.  2192067            200
#> 10                   Darkest Hour       Focus Features  1622510             73
#> 11             The Shape of Water      Fox Searchlight  1419891             95
#> 12                   Molly’s Game    STX Entertainment  1172036             NA
#> 13                         Wonder            Lionsgate   710214             77
#> 14                 Daddy’s Home 2   Paramount Pictures   544612            149
#> 15                 Justice League         Warner Bros.   435091             76
#> 16                      Lady Bird                  A24   415989             60
#> 17                 Thor: Ragnarok          Walt Disney   329457             64
#> 18            The Disaster Artist                  A24   325057             77
#> 19                       The Star        Sony Pictures   238704            -13
#> 20                       The Post     20th Century Fox   236046             35
#> 21                       I, Tonya                 Neon   210755            109
#> 22   Three Billboards Outside Eb…      Fox Searchlight   207975             42
#> 23   Murder on the Orient Express     20th Century Fox   203411             58
#> 24                 Phantom Thread       Focus Features   127272             NA
#> 25           A Bad Moms Christmas    STX Entertainment    60302            122
#> 26                   Wonder Wheel       Amazon Studios    42778             67
#> 27 The Man Who Invented Christmas      Bleecker Street    33190             48
#> 28                           Jane     Abramorama Films    27215             NA
#> 29                       Geostorm         Warner Bros.    16062            169
#> 30                 Only the Brave        Sony Pictures     9790            145
#> 31                       Hostiles Entertainment Studi…     9606             65
#> 32            The Florida Project                  A24     9430             46
#> 33              Blade Runner 2049         Warner Bros.     8433             52
#> 34          Roman J. Israel, Esq.        Sony Pictures     4786             63
#> 35   Same Kind of Different as Me Pure Flix Entertain…     3069            135
#> 36             Victoria and Abdul       Focus Features     2840            114
#> 37                      Big Sonia       Argot Pictures     2311             53
#> 38                       Marshall            Open Road     1937             83
#> 39                The Breadwinner                GKIDS      792             27
#> 40             Let There Be Light   Atlas Distribution      736              1
#> 41   The Killing of a Sacred Deer                  A24      602              3
#> 42                         Thelma          The Orchard      593             56
#> 43         BPM (Beats per Minute)          The Orchard      160            -35
#> 44   Birdboy: The Forgotten Chil…                GKIDS      117            -25
#> 45   Star Wars Ep. VIII: The Las…          Walt Disney 13566649            -32
#> 46 Jumanji: Welcome to the Jungle        Sony Pictures 12989945            -33
#> 47           The Greatest Showman     20th Century Fox  4436321            -23
#> 48                Pitch Perfect 3            Universal  3565285            -46
#> 49                      Ferdinand     20th Century Fox  2757842            -33
#> 50                           Coco          Walt Disney  1852339            -33
#> 51                   Darkest Hour       Focus Features  1771920            -10
#> 52     All the Money in the World        Sony Pictures  1749655            -16
#> 53                     Downsizing   Paramount Pictures  1273053            -31
#> 54                 Father Figures         Warner Bros.  1095601            -27
#> 55             The Shape of Water      Fox Searchlight   996207            -27
#> 56                         Wonder            Lionsgate   847181            -28
#> 57                   Molly’s Game    STX Entertainment   729998            -17
#> 58                      Lady Bird                  A24   387940            -32
#> 59                 Justice League         Warner Bros.   276222            -44
#> 60                 Daddy’s Home 2   Paramount Pictures   244693            -33
#> 61                 Thor: Ragnarok          Walt Disney   206623            -40
#> 62   Three Billboards Outside Eb…      Fox Searchlight   201312            -24
#> 63            The Disaster Artist                  A24   200119            -49
#> 64                       The Post     20th Century Fox   186640             -8
#> 65   Murder on the Orient Express     20th Century Fox   176205            -24
#> 66                       I, Tonya                 Neon   155000            -38
#> 67                       The Star        Sony Pictures    98913            -46
#> 68                 Phantom Thread       Focus Features    63340            -23
#> 69           A Bad Moms Christmas    STX Entertainment    32523            -51
#> 70 The Man Who Invented Christmas      Bleecker Street    17652            -31
#> 71            The Florida Project                  A24    11940            -46
#> 72                       Geostorm         Warner Bros.    11702            -51
#> 73                   Wonder Wheel       Amazon Studios    11088            -21
#> 74              Blade Runner 2049         Warner Bros.    10427            -48
#> 75                 Only the Brave        Sony Pictures     9765            -46
#> 76                       Hostiles Entertainment Studi…     9739            -26
#> 77             Victoria and Abdul       Focus Features     3485            -28
#> 78   Same Kind of Different as Me Pure Flix Entertain…     2919            -38
#> 79                       Marshall            Open Road     2687            -42
#> 80          Roman J. Israel, Esq.        Sony Pictures     2322            -54
#> 81                      Big Sonia       Argot Pictures     2298            -45
#> 82                The Breadwinner                GKIDS      528            -76
#> 83             Let There Be Light   Atlas Distribution      504            -53
#> 84                         Thelma          The Orchard      245            -63
#>    theaters per_theater total_gross days       date
#> 1      4232        6489   395627411   11 2017-12-25
#> 2      3765        5083    71913848    6 2017-12-25
#> 3      3447        1885    26424890    4 2017-12-25
#> 4      3006        1864    19008847    6 2017-12-25
#> 5      2111        1329   164307743   34 2017-12-25
#> 6      3630         772    29601410   11 2017-12-25
#> 7      2668        1018     7670130    4 2017-12-25
#> 8      2074        1237     2565393    1 2017-12-25
#> 9      2902         755     5479518    4 2017-12-25
#> 10      806        2013     8367532   34 2017-12-25
#> 11      730        1945     8986046   25 2017-12-25
#> 12      271        4325     1172036    1 2017-12-25
#> 13     1130         629   115654826   39 2017-12-25
#> 14     1073         508    99751582   46 2017-12-25
#> 15     1101         395   223128877   39 2017-12-25
#> 16      402        1035    28715314   53 2017-12-25
#> 17      701         470   309454093   53 2017-12-25
#> 18      517         629    16020034   25 2017-12-25
#> 19     1106         216    38594870   39 2017-12-25
#> 20        9       26227      762057    4 2017-12-25
#> 21       42        5018     1277150   18 2017-12-25
#> 22      264         788    22949085   46 2017-12-25
#> 23      361         563    99444878   46 2017-12-25
#> 24        4       31818      127272    1 2017-12-25
#> 25      275         219    71571546   55 2017-12-25
#> 26      113         379     1205919   25 2017-12-25
#> 27      146         227     5419083   34 2017-12-25
#> 28       19        1432     1432348   67 2017-12-25
#> 29      110         146    33473188   67 2017-12-25
#> 30       86         114    18106612   67 2017-12-25
#> 31        3        3202       32455    4 2017-12-25
#> 32       43         219     5204890   81 2017-12-25
#> 33       75         112    91564672   81 2017-12-25
#> 34       49          98    11825352   39 2017-12-25
#> 35       45          68     6370123   67 2017-12-25
#> 36       24         118    22205730   95 2017-12-25
#> 37        3         770       94741   39 2017-12-25
#> 38       58          33     9450507   74 2017-12-25
#> 39       10          79      210278   39 2017-12-25
#> 40        8          92     7205484   60 2017-12-25
#> 41        8          75     2289084   67 2017-12-25
#> 42        9          66      127398   46 2017-12-25
#> 43        1         160       92117   67 2017-12-25
#> 44        1         117       12620   11 2017-12-25
#> 45     4232        3206   517218368   17 2017-12-31
#> 46     3765        3450   169002557   12 2017-12-31
#> 47     3316        1338    49036718   12 2017-12-31
#> 48     3468        1028    63300605   10 2017-12-31
#> 49     3337         826    53546717   17 2017-12-31
#> 50     2104         880   179834199   40 2017-12-31
#> 51      943        1879    18132857   40 2017-12-31
#> 52     2074         844    12712380    7 2017-12-31
#> 53     2664         478    17160674   10 2017-12-31
#> 54     2902         378    12885287   10 2017-12-31
#> 55      756        1318    15670177   31 2017-12-31
#> 56     1193         710   121559294   45 2017-12-31
#> 57      271        2694     5377480    7 2017-12-31
#> 58      392         990    31432463   59 2017-12-31
#> 59     1215         227   225546295   45 2017-12-31
#> 60     1073         228   101791381   52 2017-12-31
#> 61      540         383   311225150   59 2017-12-31
#> 62      266         757    24164629   52 2017-12-31
#> 63      507         395    17861795   31 2017-12-31
#> 64        9       20738     1760121   10 2017-12-31
#> 65      391         451   100619624   52 2017-12-31
#> 66       49        3163     2351336   24 2017-12-31
#> 67      661         150    39963949   45 2017-12-31
#> 68        4       15835      527840    7 2017-12-31
#> 69      214         152    71891988   61 2017-12-31
#> 70      146         121     5575759   40 2017-12-31
#> 71       33         362     5297299   87 2017-12-31
#> 72      110         106    33578782   73 2017-12-31
#> 73       55         202     1306606   31 2017-12-31
#> 74       75         139    91654512   87 2017-12-31
#> 75       86         114    18181288   73 2017-12-31
#> 76        5        1948       85745   10 2017-12-31
#> 77       23         152    22233285  101 2017-12-31
#> 78       40          73     6394006   73 2017-12-31
#> 79       50          54     9468318   80 2017-12-31
#> 80       33          70    11851274   45 2017-12-31
#> 81        1        2298      113604   45 2017-12-31
#> 82       10          53      220307   45 2017-12-31
#> 83        8          63     7210068   66 2017-12-31
#> 84        6          41      132011   52 2017-12-31