Get the top N ranking grossing movies

Get the top N ranking grossing movies

top_grossing(type = "american", ranks = 1:100)

Arguments

type
A string that says which type of box office sorted ranking you want. Options are 'american' (American box office), 'international' (non-American) and 'worldwide' (domestic + international box office).

ranks
A vector of rankings you want it to return. For example. an input of 1:5 will return the top 5 grossing movies.

Value

Data frame returning info on the name of the movie, it's rank, the year the movie was released, and the total gross from domestic (American), international, and total ticket sales.

Examples

top_grossing()

#> Please note that these numbers are not adjusted for inflation.

#>     rank                                              movie year_released
#> 2      1               Star Wars Ep. VII: The Force Awakens          2015
#> 3      2                                  Avengers: Endgame          2019
#> 4      3                                             Avatar          2009
#> 5      4                                      Black Panther          2018
#> 6      5                             Avengers: Infinity War          2018
#> 7      6                                            Titanic          1997
#> 8      7                                     Jurassic World          2015
#> 9      8                                       The Avengers          2012
#> 10     9                  Star Wars Ep. VIII: The Last Jedi          2017
#> 11    10                                      Incredibles 2          2018
#> 12    11                                      The Lion King          2019
#> 13    12                                    The Dark Knight          2008
#> 14    13                       Rogue One: A Star Wars Story          2016
#> 15    14                               Beauty and the Beast          2017
#> 16    15                                       Finding Dory          2016
#> 17    16                Star Wars Ep. I: The Phantom Menace          1999
#> 18    17                       Star Wars Ep. IV: A New Hope          1977
#> 19    18                            Avengers: Age of Ultron          2015
#> 20    19                              The Dark Knight Rises          2012
#> 21    20                                            Shrek 2          2004
#> 22    21                          ET: The Extra-Terrestrial          1982
#> 23    22                                        Toy Story 4          2019
#> 24    23                                     Captain Marvel          2019
#> 25    24                    The Hunger Games: Catching Fire          2013
#> 26    25         Pirates of the Caribbean: Dead Man’s Chest          2006
#> 27    26                                      The Lion King          1994
#> 28    27                     Jurassic World: Fallen Kingdom          2018
#> 29    28                                        Toy Story 3          2010
#> 30    29                                       Wonder Woman          2017
#> 31    30                                         Iron Man 3          2013
#> 32    31                         Captain America: Civil War          2016
#> 33    32                                   The Hunger Games          2012
#> 34    33                     Jumanji: Welcome to the Jungle          2017
#> 35    34                                         Spider-Man          2002
#> 36    35                                      Jurassic Park          1993
#> 37    36                Transformers: Revenge of the Fallen          2009
#> 38    37                                             Frozen          2013
#> 39    38                          Spider-Man: Far From Home          2019
#> 40    39                      Guardians of the Galaxy Vol 2          2017
#> 41    40      Harry Potter and the Deathly Hallows: Part II          2011
#> 42    41                                       Finding Nemo          2003
#> 43    42             Star Wars Ep. III: Revenge of the Sith          2005
#> 44    43      The Lord of the Rings: The Return of the King          2003
#> 45    44                                       Spider-Man 2          2004
#> 46    45                          The Passion of the Christ          2004
#> 47    46                            The Secret Life of Pets          2016
#> 48    47                                    Despicable Me 2          2013
#> 49    48                                    The Jungle Book          2016
#> 50    49                                           Deadpool          2016
#> 51    50                                         Inside Out          2015
#> 52    51                                            Aladdin          2019
#> 53    52                                          Furious 7          2015
#> 54    53                     Transformers: Dark of the Moon          2011
#> 55    54                                    American Sniper          2014
#> 56    55              The Lord of the Rings: The Two Towers          2002
#> 57    56                                           Zootopia          2016
#> 58    57              The Hunger Games: Mockingjay - Part 1          2014
#> 59    58                                       Spider-Man 3          2007
#> 60    59                                            Minions          2015
#> 61    60                                            Aquaman          2018
#> 62    61                             Spider-Man: Homecoming          2017
#> 63    62                                Alice in Wonderland          2010
#> 64    63                            Guardians of the Galaxy          2014
#> 65    64                 Batman v Superman: Dawn of Justice          2016
#> 66    65                                       Forrest Gump          1994
#> 67    66                                                 It          2017
#> 68    67                                      Suicide Squad          2016
#> 69    68                                         Deadpool 2          2018
#> 70    69                                    Shrek the Third          2007
#> 71    70                                       Transformers          2007
#> 72    71                                           Iron Man          2008
#> 73    72              Harry Potter and the Sorcerer’s Stone          2001
#> 74    73 Indiana Jones and the Kingdom of the Crystal Skull          2008
#> 75    74  The Lord of the Rings: The Fellowship of the Ring          2001
#> 76    75                                     Thor: Ragnarok          2017
#> 77    76                                         Iron Man 2          2010
#> 78    77             Star Wars Ep. II: Attack of the Clones          2002
#> 79    78           Pirates of the Caribbean: At World’s End          2007
#> 80    79               Star Wars Ep. VI: Return of the Jedi          1983
#> 81    80                                   Independence Day          1996
#> 82    81   Pirates of the Caribbean: The Curse of the Blac…          2003
#> 83    82                                            Skyfall          2012
#> 84    83                  The Hobbit: An Unexpected Journey          2012
#> 85    84             Harry Potter and the Half-Blood Prince          2009
#> 86    85                         The Twilight Saga: Eclipse          2010
#> 87    86                        The Twilight Saga: New Moon          2009
#> 88    87       Harry Potter and the Deathly Hallows: Part I          2010
#> 89    88                                    The Sixth Sense          1999
#> 90    89                                                 Up          2009
#> 91    90                                          Inception          2010
#> 92    91           The Twilight Saga: Breaking Dawn, Part 2          2012
#> 93    92          Harry Potter and the Order of the Phoenix          2007
#> 94    93   The Chronicles of Narnia: The Lion, the Witch a…          2005
#> 95    94                                       Man of Steel          2013
#> 96    95           Star Wars Ep. V: The Empire Strikes Back          1980
#> 97    96                Harry Potter and the Goblet of Fire          2005
#> 98    97                                     Monsters, Inc.          2001
#> 99    98                                         Home Alone          1990
#> 100   99              The Hunger Games: Mockingjay - Part 2          2015
#> 101  100                                The Matrix Reloaded          2003
#>     american_box_office international_box_office total_box_office
#> 2             936662225               1116648995       2053311220
#> 3             858373000               1937100000       2795473000
#> 4             760507625               2029197650       2789705275
#> 5             700059566                648198658       1348258224
#> 6             678815482               1369318718       2048134200
#> 7             659363944               1548844451       2208208395
#> 8             652270625                996584239       1648854864
#> 9             623279547                894656350       1517935897
#> 10            620181382                696540365       1316721747
#> 11            608581744                633926020       1242507764
#> 12            543041133               1098700000       1641741133
#> 13            533720947                468275260       1001996207
#> 14            532177324                516925532       1049102856
#> 15            504014165                755185541       1259199706
#> 16            486295561                534919632       1021215193
#> 17            474544677                552500000       1027044677
#> 18            460998007                325600000        786598007
#> 19            459005868                944008095       1403013963
#> 20            448139099                636300000       1084439099
#> 21            441226247                495781885        937008132
#> 22            435110554                357854772        792965326
#> 23            433742240                625500000       1059242240
#> 24            426829839                699300000       1126129839
#> 25            424668047                440200000        864868047
#> 26            423315812                642900000       1066215812
#> 27            421785283                564429585        986214868
#> 28            417719760                888053039       1305772799
#> 29            415004880                653874642       1068879522
#> 30            412563408                408569970        821133378
#> 31            408992272                806400000       1215392272
#> 32            408084349                731985064       1140069413
#> 33            408010692                269912687        677923379
#> 34            404508916                559987277        964496193
#> 35            403706375                418000000        821706375
#> 36            402523348                643104279       1045627627
#> 37            402111870                434407829        836519699
#> 38            400738009                871731901       1272469910
#> 39            390532085                741258837       1131790922
#> 40            389813101                472503132        862316233
#> 41            381193157                960500000       1341693157
#> 42            380529370                555900000        936429370
#> 43            380270577                468728300        848998877
#> 44            377845905                763557436       1141403341
#> 45            373524485                421586185        795110670
#> 46            370782930                251558994        622341924
#> 47            368384330                518222082        886606412
#> 48            368065385                607151450        975216835
#> 49            364001123                598853424        962854547
#> 50            363070709                437954884        801025593
#> 51            356461711                497774281        854235992
#> 52            355559216                682500000       1038059216
#> 53            353007020               1165715774       1518722794
#> 54            352390543                771400000       1123790543
#> 55            350126372                197200000        547326372
#> 56            342548984                592150661        934699645
#> 57            341268248                678161368       1019429616
#> 58            337135885                429439246        766575131
#> 59            336530303                558329927        894860230
#> 60            336045770                824290403       1160336173
#> 61            335061807                811832833       1146894640
#> 62            334201140                545965210        880166350
#> 63            334191110                691300000       1025491110
#> 64            333172112                437695404        770867516
#> 65            330360194                537140087        867500281
#> 66            330151138                349699499        679850637
#> 67            327481748                372900000        700381748
#> 68            325100054                420959833        746059887
#> 69            324591735                462088822        786680557
#> 70            322719944                484610992        807330936
#> 71            319246193                389026399        708272592
#> 72            318604126                266567421        585171547
#> 73            317871467                657176139        975047606
#> 74            317101119                469534294        786635413
#> 75            315544750                571666235        887210985
#> 76            315058289                531921735        846980024
#> 77            312433331                308723058        621156389
#> 78            310676740                346018875        656695615
#> 79            309420425                654000000        963420425
#> 80            309205079                263500000        572705079
#> 81            306169255                511231623        817400878
#> 82            305411224                329542879        634954103
#> 83            304360277                806166704       1110526981
#> 84            303003568                714000000       1017003568
#> 85            302089278                633124489        935213767
#> 86            300531751                405571077        706102828
#> 87            296623634                390934093        687557727
#> 88            296131568                664300000        960431568
#> 89            293506292                379300000        672806292
#> 90            293004164                438459213        731463377
#> 91            292576195                542948447        835524642
#> 92            292324737                537400000        829724737
#> 93            292137260                650939197        943076457
#> 94            291710957                428828615        720539572
#> 95            291045518                376954000        667999518
#> 96            290271960                243889374        534161334
#> 97            290201752                606898042        897099794
#> 98            289423425                270334294        559757719
#> 99            285761243                190923432        476684675
#> 100           281723902                367262885        648986787
#> 101           281553689                457023240        738576929

top_grossing(ranks = 1:5)

#> Please note that these numbers are not adjusted for inflation.

#>   rank                                movie year_released american_box_office
#> 2    1 Star Wars Ep. VII: The Force Awakens          2015           936662225
#> 3    2                    Avengers: Endgame          2019           858373000
#> 4    3                               Avatar          2009           760507625
#> 5    4                        Black Panther          2018           700059566
#> 6    5               Avengers: Infinity War          2018           678815482
#>   international_box_office total_box_office
#> 2               1116648995       2053311220
#> 3               1937100000       2795473000
#> 4               2029197650       2789705275
#> 5                648198658       1348258224
#> 6               1369318718       2048134200

top_grossing(type = "international")

#> Please note that these numbers are not adjusted for inflation.

#>     rank                                             movie year_released
#> 2      1                                            Avatar          2009
#> 3      2                                 Avengers: Endgame          2019
#> 4      3                                           Titanic          1997
#> 5      4                            Avengers: Infinity War          2018
#> 6      5                                         Furious 7          2015
#> 7      6              Star Wars Ep. VII: The Force Awakens          2015
#> 8      7                                     The Lion King          2019
#> 9      8                           The Fate of the Furious          2017
#> 10     9                                    Jurassic World          2015
#> 11    10     Harry Potter and the Deathly Hallows: Part II          2011
#> 12    11                           Avengers: Age of Ultron          2015
#> 13    12                                      The Avengers          2012
#> 14    13                    Jurassic World: Fallen Kingdom          2018
#> 15    14                                            Frozen          2013
#> 16    15                   Transformers: Age of Extinction          2014
#> 17    16                                       Zhan Lang 2          2017
#> 18    17                                           Minions          2015
#> 19    18                                           Aquaman          2018
#> 20    19                                        Iron Man 3          2013
#> 21    20                                           Skyfall          2012
#> 22    21       Pirates of the Caribbean: On Stranger Tides          2011
#> 23    22                    Transformers: Dark of the Moon          2011
#> 24    23                                   Despicable Me 3          2017
#> 25    24     The Lord of the Rings: The Return of the King          2003
#> 26    25                              Beauty and the Beast          2017
#> 27    26                         Spider-Man: Far From Home          2019
#> 28    27                        Captain America: Civil War          2016
#> 29    28                      Ne Zha Zhi Mo Tong Jiang Shi          2019
#> 30    29                        Ice Age: Continental Drift          2012
#> 31    30                 The Hobbit: An Unexpected Journey          2012
#> 32    31               The Hobbit: The Desolation of Smaug          2013
#> 33    32                                    Captain Marvel          2019
#> 34    33                 Star Wars Ep. VIII: The Last Jedi          2017
#> 35    34                                   Liu Lang Di Qiu          2019
#> 36    35                               Alice in Wonderland          2010
#> 37    36         The Hobbit: The Battle of the Five Armies          2014
#> 38    37                                           Aladdin          2019
#> 39    38                                           Spectre          2015
#> 40    39                                          Zootopia          2016
#> 41    40                                 Bohemian Rhapsody          2018
#> 42    41      Harry Potter and the Deathly Hallows: Part I          2010
#> 43    42                    Ice Age: Dawn of the Dinosaurs          2009
#> 44    43             Harry Potter and the Sorcerer’s Stone          2001
#> 45    44          Pirates of the Caribbean: At World’s End          2007
#> 46    45                                       Toy Story 3          2010
#> 47    46         Harry Potter and the Order of the Phoenix          2007
#> 48    47                                     Black Panther          2018
#> 49    48                                     Jurassic Park          1993
#> 50    49        Pirates of the Caribbean: Dead Man’s Chest          2006
#> 51    50                                             Venom          2018
#> 52    51                             The Dark Knight Rises          2012
#> 53    52                                     Incredibles 2          2018
#> 54    53            Harry Potter and the Half-Blood Prince          2009
#> 55    54                                       Toy Story 4          2019
#> 56    55           Harry Potter and the Chamber of Secrets          2002
#> 57    56  Pirates of the Caribbean: Dead Men Tell No Tales          2017
#> 58    57                                   Despicable Me 2          2013
#> 59    58               Harry Potter and the Goblet of Fire          2005
#> 60    59                                   The Jungle Book          2016
#> 61    60             The Lord of the Rings: The Two Towers          2002
#> 62    61                                              2012          2009
#> 63    62                                              Coco          2017
#> 64    63             Fast & Furious Presents: Hobbs & Shaw          2019
#> 65    64                                Hong Hai Xing Dong          2018
#> 66    65 The Lord of the Rings: The Fellowship of the Ring          2001
#> 67    66           Fantastic Beasts and Where to Find Them          2016
#> 68    67                       Mission: Impossible—Fallout          2018
#> 69    68                                     The Lion King          1994
#> 70    69                    Jumanji: Welcome to the Jungle          2017
#> 71    70                                      Spider-Man 3          2007
#> 72    71                                      Finding Nemo          2003
#> 73    72               Star Wars Ep. I: The Phantom Menace          1999
#> 74    73                                        Mei Ren Yu          2016
#> 75    74                                Fast and Furious 6          2013
#> 76    75                                 The Da Vinci Code          2006
#> 77    76          Harry Potter and the Prisoner of Azkaban          2004
#> 78    77                            Spider-Man: Homecoming          2017
#> 79    78                                         Inception          2010
#> 80    79                             Tang Ren Jie Tan An 2          2018
#> 81    80          The Twilight Saga: Breaking Dawn, Part 2          2012
#> 82    81                Batman v Superman: Dawn of Justice          2016
#> 83    82                                      Finding Dory          2016
#> 84    83                                    Thor: Ragnarok          2017
#> 85    84                Madagascar 3: Europe's Most Wanted          2012
#> 86    85                           The Secret Life of Pets          2016
#> 87    86                               Shrek Forever After          2010
#> 88    87                                        Maleficent          2014
#> 89    88                      Rogue One: A Star Wars Story          2016
#> 90    89                        X-Men: Days of Future Past          2014
#> 91    90                                  Independence Day          1996
#> 92    91                          The Amazing Spider-Man 2          2014
#> 93    92                    Dawn of the Planet of the Apes          2014
#> 94    93                                   Kung Fu Panda 2          2011
#> 95    94                                        Inside Out          2015
#> 96    95                                        Life of Pi          2012
#> 97    96                            The Amazing Spider-Man          2012
#> 98    97                                           Shrek 2          2004
#> 99    98                  Mission: Impossible—Rogue Nation          2015
#> 100   99       Fantastic Beasts: The Crimes of Grindelwald          2018
#> 101  100                                             Joker          2019
#>     american_box_office international_box_office total_box_office
#> 2             760507625               2029197650       2789705275
#> 3             858373000               1937100000       2795473000
#> 4             659363944               1548844451       2208208395
#> 5             678815482               1369318718       2048134200
#> 6             353007020               1165715774       1518722794
#> 7             936662225               1116648995       2053311220
#> 8             543041133               1098700000       1641741133
#> 9             225764765               1009081502       1234846267
#> 10            652270625                996584239       1648854864
#> 11            381193157                960500000       1341693157
#> 12            459005868                944008095       1403013963
#> 13            623279547                894656350       1517935897
#> 14            417719760                888053039       1305772799
#> 15            400738009                871731901       1272469910
#> 16            245439076                858600000       1104039076
#> 17              2721100                855881967        858603067
#> 18            336045770                824290403       1160336173
#> 19            335061807                811832833       1146894640
#> 20            408992272                806400000       1215392272
#> 21            304360277                806166704       1110526981
#> 22            241063875                804600000       1045663875
#> 23            352390543                771400000       1123790543
#> 24            264624300                770100410       1034724710
#> 25            377845905                763557436       1141403341
#> 26            504014165                755185541       1259199706
#> 27            390532085                741258837       1131790922
#> 28            408084349                731985064       1140069413
#> 29              3695533                725297824        728993357
#> 30            161321843                718443294        879765137
#> 31            303003568                714000000       1017003568
#> 32            258366855                702000000        960366855
#> 33            426829839                699300000       1126129839
#> 34            620181382                696540365       1316721747
#> 35                   NA                692163684        692163684
#> 36            334191110                691300000       1025491110
#> 37            255119788                690457833        945577621
#> 38            355559216                682500000       1038059216
#> 39            200074175                679546748        879620923
#> 40            341268248                678161368       1019429616
#> 41            216303339                676714322        893017661
#> 42            296131568                664300000        960431568
#> 43            196573705                663128152        859701857
#> 44            317871467                657176139        975047606
#> 45            309420425                654000000        963420425
#> 46            415004880                653874642       1068879522
#> 47            292137260                650939197        943076457
#> 48            700059566                648198658       1348258224
#> 49            402523348                643104279       1045627627
#> 50            423315812                642900000       1066215812
#> 51            213511408                640117197        853628605
#> 52            448139099                636300000       1084439099
#> 53            608581744                633926020       1242507764
#> 54            302089278                633124489        935213767
#> 55            433742240                625500000       1059242240
#> 56            262233381                616991754        879225135
#> 57            172558876                615682261        788241137
#> 58            368065385                607151450        975216835
#> 59            290201752                606898042        897099794
#> 60            364001123                598853424        962854547
#> 61            342548984                592150661        934699645
#> 62            166112167                591565581        757677748
#> 63            209726015                588275967        798001982
#> 64            173732585                586366411        760098996
#> 65              1543547                576647204        578190751
#> 66            315544750                571666235        887210985
#> 67            234037575                568365278        802402853
#> 68            220159104                567297448        787456552
#> 69            421785283                564429585        986214868
#> 70            404508916                559987277        964496193
#> 71            336530303                558329927        894860230
#> 72            380529370                555900000        936429370
#> 73            474544677                552500000       1027044677
#> 74              3229457                551287214        554516671
#> 75            238679850                550620594        789300444
#> 76            217536138                550284321        767820459
#> 77            249757726                547149597        796907323
#> 78            334201140                545965210        880166350
#> 79            292576195                542948447        835524642
#> 80              1983984                542084590        544068574
#> 81            292324737                537400000        829724737
#> 82            330360194                537140087        867500281
#> 83            486295561                534919632       1021215193
#> 84            315058289                531921735        846980024
#> 85            216391482                530529789        746921271
#> 86            368384330                518222082        886606412
#> 87            238736787                517507886        756244673
#> 88            241407328                517129407        758536735
#> 89            532177324                516925532       1049102856
#> 90            233921534                513941241        747862775
#> 91            306169255                511231623        817400878
#> 92            202853933                506142403        708996336
#> 93            208545589                502098977        710644566
#> 94            165249063                499588484        664837547
#> 95            356461711                497774281        854235992
#> 96            124987022                495924981        620912003
#> 97            262030663                495859604        757890267
#> 98            441226247                495781885        937008132
#> 99            195042377                493816615        688858992
#> 100           159555901                492664185        652220086
#> 101           258683522                490996649        749680171

top_grossing(type = "international", ranks = 1:10)

#> Please note that these numbers are not adjusted for inflation.

#>    rank                                         movie year_released
#> 2     1                                        Avatar          2009
#> 3     2                             Avengers: Endgame          2019
#> 4     3                                       Titanic          1997
#> 5     4                        Avengers: Infinity War          2018
#> 6     5                                     Furious 7          2015
#> 7     6          Star Wars Ep. VII: The Force Awakens          2015
#> 8     7                                 The Lion King          2019
#> 9     8                       The Fate of the Furious          2017
#> 10    9                                Jurassic World          2015
#> 11   10 Harry Potter and the Deathly Hallows: Part II          2011
#>    american_box_office international_box_office total_box_office
#> 2            760507625               2029197650       2789705275
#> 3            858373000               1937100000       2795473000
#> 4            659363944               1548844451       2208208395
#> 5            678815482               1369318718       2048134200
#> 6            353007020               1165715774       1518722794
#> 7            936662225               1116648995       2053311220
#> 8            543041133               1098700000       1641741133
#> 9            225764765               1009081502       1234846267
#> 10           652270625                996584239       1648854864
#> 11           381193157                960500000       1341693157