Get the top N ranking grossing movies

Get the top N ranking grossing movies

top_grossing(type = "american", ranks = 1:100)

Arguments

type
A string that says which type of box office sorted ranking you want. Options are 'american' (American box office), 'international' (non-American) and 'worldwide' (domestic + international box office).

ranks
A vector of rankings you want it to return. For example. an input of 1:5 will return the top 5 grossing movies.

Value

Data frame returning info on the name of the movie, it's rank, the year the movie was released, and the total gross from domestic (American), international, and total ticket sales.

Examples

top_grossing()

#> Please note that these numbers are not adjusted for inflation.

#>     rank                                              movie year_released
#> 2      1               Star Wars Ep. VII: The Force Awakens          2015
#> 3      2                                  Avengers: Endgame          2019
#> 4      3                                             Avatar          2009
#> 5      4                                      Black Panther          2018
#> 6      5                             Avengers: Infinity War          2018
#> 7      6                                            Titanic          1997
#> 8      7                                     Jurassic World          2015
#> 9      8                                       The Avengers          2012
#> 10     9                  Star Wars Ep. VIII: The Last Jedi          2017
#> 11    10                                      Incredibles 2          2018
#> 12    11                                      The Lion King          2019
#> 13    12                                    The Dark Knight          2008
#> 14    13                       Rogue One: A Star Wars Story          2016
#> 15    14                               Beauty and the Beast          2017
#> 16    15                                       Finding Dory          2016
#> 17    16                Star Wars Ep. I: The Phantom Menace          1999
#> 18    17                       Star Wars Ep. IV: A New Hope          1977
#> 19    18                            Avengers: Age of Ultron          2015
#> 20    19                              The Dark Knight Rises          2012
#> 21    20                                            Shrek 2          2004
#> 22    21                          ET: The Extra-Terrestrial          1982
#> 23    22                                        Toy Story 4          2019
#> 24    23                                     Captain Marvel          2019
#> 25    24                    The Hunger Games: Catching Fire          2013
#> 26    25         Pirates of the Caribbean: Dead Man’s Chest          2006
#> 27    26                                      The Lion King          1994
#> 28    27                     Jurassic World: Fallen Kingdom          2018
#> 29    28                                        Toy Story 3          2010
#> 30    29                                       Wonder Woman          2017
#> 31    30                                         Iron Man 3          2013
#> 32    31                         Captain America: Civil War          2016
#> 33    32                                   The Hunger Games          2012
#> 34    33                                         Spider-Man          2002
#> 35    34                     Jumanji: Welcome to the Jungle          2017
#> 36    35                                      Jurassic Park          1993
#> 37    36                Transformers: Revenge of the Fallen          2009
#> 38    37                                             Frozen          2013
#> 39    38                          Spider-Man: Far From Home          2019
#> 40    39                      Guardians of the Galaxy Vol 2          2017
#> 41    40      Harry Potter and the Deathly Hallows: Part II          2011
#> 42    41                                       Finding Nemo          2003
#> 43    42             Star Wars Ep. III: Revenge of the Sith          2005
#> 44    43      The Lord of the Rings: The Return of the King          2003
#> 45    44                                       Spider-Man 2          2004
#> 46    45                          The Passion of the Christ          2004
#> 47    46                            The Secret Life of Pets          2016
#> 48    47                                    Despicable Me 2          2013
#> 49    48                                    The Jungle Book          2016
#> 50    49                                           Deadpool          2016
#> 51    50                                         Inside Out          2015
#> 52    51                                            Aladdin          2019
#> 53    52                                          Furious 7          2015
#> 54    53                     Transformers: Dark of the Moon          2011
#> 55    54                                    American Sniper          2014
#> 56    55              The Lord of the Rings: The Two Towers          2002
#> 57    56                                           Zootopia          2016
#> 58    57              The Hunger Games: Mockingjay - Part 1          2014
#> 59    58                                       Spider-Man 3          2007
#> 60    59                                            Minions          2015
#> 61    60                                            Aquaman          2018
#> 62    61                             Spider-Man: Homecoming          2017
#> 63    62                                Alice in Wonderland          2010
#> 64    63                            Guardians of the Galaxy          2014
#> 65    64                 Batman v Superman: Dawn of Justice          2016
#> 66    65                                       Forrest Gump          1994
#> 67    66                                              Joker          2019
#> 68    67                                                 It          2017
#> 69    68                                      Suicide Squad          2016
#> 70    69                                         Deadpool 2          2018
#> 71    70                                    Shrek the Third          2007
#> 72    71                                       Transformers          2007
#> 73    72                                           Iron Man          2008
#> 74    73              Harry Potter and the Sorcerer’s Stone          2001
#> 75    74 Indiana Jones and the Kingdom of the Crystal Skull          2008
#> 76    75  The Lord of the Rings: The Fellowship of the Ring          2001
#> 77    76                                     Thor: Ragnarok          2017
#> 78    77                                         Iron Man 2          2010
#> 79    78             Star Wars Ep. II: Attack of the Clones          2002
#> 80    79           Pirates of the Caribbean: At World’s End          2007
#> 81    80               Star Wars Ep. VI: Return of the Jedi          1983
#> 82    81                                   Independence Day          1996
#> 83    82   Pirates of the Caribbean: The Curse of the Blac…          2003
#> 84    83                                            Skyfall          2012
#> 85    84                  The Hobbit: An Unexpected Journey          2012
#> 86    85             Harry Potter and the Half-Blood Prince          2009
#> 87    86                         The Twilight Saga: Eclipse          2010
#> 88    87                        The Twilight Saga: New Moon          2009
#> 89    88       Harry Potter and the Deathly Hallows: Part I          2010
#> 90    89                                    The Sixth Sense          1999
#> 91    90                                                 Up          2009
#> 92    91                                          Inception          2010
#> 93    92           The Twilight Saga: Breaking Dawn, Part 2          2012
#> 94    93          Harry Potter and the Order of the Phoenix          2007
#> 95    94   The Chronicles of Narnia: The Lion, the Witch a…          2005
#> 96    95                                       Man of Steel          2013
#> 97    96           Star Wars Ep. V: The Empire Strikes Back          1980
#> 98    97                Harry Potter and the Goblet of Fire          2005
#> 99    98                                     Monsters, Inc.          2001
#> 100   99                                         Home Alone          1990
#> 101  100              The Hunger Games: Mockingjay - Part 2          2015
#>     american_box_office international_box_office total_box_office
#> 2             936662225               1131561399       2068223624
#> 3             858373000               1939427564       2797800564
#> 4             760507625               2029197650       2789705275
#> 5             700059566                646853595       1346913161
#> 6             678815482               1369544272       2048359754
#> 7             659363944               1548844451       2208208395
#> 8             652270625                996584239       1648854864
#> 9             623357910                895455078       1518812988
#> 10            620181382                712358507       1332539889
#> 11            608581744                634223615       1242805359
#> 12            543617567               1111810064       1655427631
#> 13            533720947                468275260       1001996207
#> 14            532177324                523879949       1056057273
#> 15            504014165                759506961       1263521126
#> 16            486295561                542275328       1028570889
#> 17            474544677                552500000       1027044677
#> 18            460998007                325600000        786598007
#> 19            459005868                944008095       1403013963
#> 20            448139099                636300000       1084439099
#> 21            441226247                495781885        937008132
#> 22            435110554                357854772        792965326
#> 23            434022296                639356585       1073378881
#> 24            426829839                701444955       1128274794
#> 25            424668047                440200000        864868047
#> 26            423315812                642900000       1066215812
#> 27            421785283                564429585        986214868
#> 28            417719760                888053039       1305772799
#> 29            415004880                653874642       1068879522
#> 30            412563408                408569970        821133378
#> 31            408992272                806400000       1215392272
#> 32            408084349                745200000       1153284349
#> 33            408010692                269912687        677923379
#> 34            407677927                418000000        825677927
#> 35            404508916                559987277        964496193
#> 36            402523348                643104279       1045627627
#> 37            402111870                434407829        836519699
#> 38            400738009                871731901       1272469910
#> 39            390532085                741395911       1131927996
#> 40            389813101                479300000        869113101
#> 41            381193157                960500000       1341693157
#> 42            380529370                555900000        936429370
#> 43            380270577                468728300        848998877
#> 44            377845905                763557436       1141403341
#> 45            373524485                421586185        795110670
#> 46            370782930                251558994        622341924
#> 47            368384330                518222082        886606412
#> 48            368065385                607151450        975216835
#> 49            364001123                598853424        962854547
#> 50            363070709                437954884        801025593
#> 51            356461711                501100000        857561711
#> 52            355559216                695134737       1050693953
#> 53            353007020               1165715774       1518722794
#> 54            352390543                771403536       1123794079
#> 55            350126372                197200000        547326372
#> 56            342548984                592150661        934699645
#> 57            341268248                695100000       1036368248
#> 58            337135885                429439246        766575131
#> 59            336530303                558329927        894860230
#> 60            336045770                824290403       1160336173
#> 61            335061807                811832833       1146894640
#> 62            334201140                545965210        880166350
#> 63            334191110                691300000       1025491110
#> 64            333172112                439600000        772772112
#> 65            330360194                537140087        867500281
#> 66            330151138                349699499        679850637
#> 67            328571522                709986847       1038558369
#> 68            327481748                372900000        700381748
#> 69            325100054                420959833        746059887
#> 70            324591735                462088822        786680557
#> 71            322719944                484610992        807330936
#> 72            319246193                389026399        708272592
#> 73            318604126                266567421        585171547
#> 74            317871467                657176139        975047606
#> 75            317101119                469534294        786635413
#> 76            315544750                571666235        887210985
#> 77            315058289                538900000        853958289
#> 78            312433331                308723058        621156389
#> 79            310676740                346018875        656695615
#> 80            309420425                654000000        963420425
#> 81            309205079                263500000        572705079
#> 82            306169255                511231623        817400878
#> 83            305410819                348900000        654310819
#> 84            304360277                806166704       1110526981
#> 85            303003568                714000000       1017003568
#> 86            302089278                633124489        935213767
#> 87            300531751                405571077        706102828
#> 88            296623634                390934093        687557727
#> 89            296131568                664300000        960431568
#> 90            293506292                379300000        672806292
#> 91            293004164                438459213        731463377
#> 92            292576195                542948447        835524642
#> 93            292324737                537400000        829724737
#> 94            292137260                650939197        943076457
#> 95            291710957                428828615        720539572
#> 96            291045518                376954000        667999518
#> 97            290271960                243889374        534161334
#> 98            290201752                606898042        897099794
#> 99            289423425                270334294        559757719
#> 100           285761243                190923432        476684675
#> 101           281723902                367262885        648986787

top_grossing(ranks = 1:5)

#> Please note that these numbers are not adjusted for inflation.

#>   rank                                movie year_released american_box_office
#> 2    1 Star Wars Ep. VII: The Force Awakens          2015           936662225
#> 3    2                    Avengers: Endgame          2019           858373000
#> 4    3                               Avatar          2009           760507625
#> 5    4                        Black Panther          2018           700059566
#> 6    5               Avengers: Infinity War          2018           678815482
#>   international_box_office total_box_office
#> 2               1131561399       2068223624
#> 3               1939427564       2797800564
#> 4               2029197650       2789705275
#> 5                646853595       1346913161
#> 6               1369544272       2048359754

top_grossing(type = "international")

#> Please note that these numbers are not adjusted for inflation.

#>     rank                                             movie year_released
#> 2      1                                            Avatar          2009
#> 3      2                                 Avengers: Endgame          2019
#> 4      3                                           Titanic          1997
#> 5      4                            Avengers: Infinity War          2018
#> 6      5                                         Furious 7          2015
#> 7      6              Star Wars Ep. VII: The Force Awakens          2015
#> 8      7                                     The Lion King          2019
#> 9      8                           The Fate of the Furious          2017
#> 10     9                                    Jurassic World          2015
#> 11    10     Harry Potter and the Deathly Hallows: Part II          2011
#> 12    11                           Avengers: Age of Ultron          2015
#> 13    12                                      The Avengers          2012
#> 14    13                    Jurassic World: Fallen Kingdom          2018
#> 15    14                                            Frozen          2013
#> 16    15                   Transformers: Age of Extinction          2014
#> 17    16                                       Zhan Lang 2          2017
#> 18    17                                           Minions          2015
#> 19    18                                           Aquaman          2018
#> 20    19                                        Iron Man 3          2013
#> 21    20                                           Skyfall          2012
#> 22    21       Pirates of the Caribbean: On Stranger Tides          2011
#> 23    22                    Transformers: Dark of the Moon          2011
#> 24    23                                   Despicable Me 3          2017
#> 25    24     The Lord of the Rings: The Return of the King          2003
#> 26    25                              Beauty and the Beast          2017
#> 27    26                        Captain America: Civil War          2016
#> 28    27                         Spider-Man: Far From Home          2019
#> 29    28                        Ice Age: Continental Drift          2012
#> 30    29                 The Hobbit: An Unexpected Journey          2012
#> 31    30                 Star Wars Ep. VIII: The Last Jedi          2017
#> 32    31                                             Joker          2019
#> 33    32                      Ne Zha Zhi Mo Tong Jiang Shi          2019
#> 34    33               The Hobbit: The Desolation of Smaug          2013
#> 35    34                                    Captain Marvel          2019
#> 36    35                                           Aladdin          2019
#> 37    36                                          Zootopia          2016
#> 38    37                                   Liu Lang Di Qiu          2019
#> 39    38                               Alice in Wonderland          2010
#> 40    39         The Hobbit: The Battle of the Five Armies          2014
#> 41    40                                 Bohemian Rhapsody          2018
#> 42    41                                           Spectre          2015
#> 43    42      Harry Potter and the Deathly Hallows: Part I          2010
#> 44    43                    Ice Age: Dawn of the Dinosaurs          2009
#> 45    44             Harry Potter and the Sorcerer’s Stone          2001
#> 46    45          Pirates of the Caribbean: At World’s End          2007
#> 47    46                                       Toy Story 3          2010
#> 48    47         Harry Potter and the Order of the Phoenix          2007
#> 49    48                                     Black Panther          2018
#> 50    49                                     Jurassic Park          1993
#> 51    50        Pirates of the Caribbean: Dead Man’s Chest          2006
#> 52    51                                             Venom          2018
#> 53    52                                       Toy Story 4          2019
#> 54    53                             The Dark Knight Rises          2012
#> 55    54                                     Incredibles 2          2018
#> 56    55            Harry Potter and the Half-Blood Prince          2009
#> 57    56           Harry Potter and the Chamber of Secrets          2002
#> 58    57  Pirates of the Caribbean: Dead Men Tell No Tales          2017
#> 59    58                                   Despicable Me 2          2013
#> 60    59               Harry Potter and the Goblet of Fire          2005
#> 61    60                                   The Jungle Book          2016
#> 62    61                                              Coco          2017
#> 63    62             The Lord of the Rings: The Two Towers          2002
#> 64    63                                              2012          2009
#> 65    64             Fast & Furious Presents: Hobbs & Shaw          2019
#> 66    65                                Hong Hai Xing Dong          2018
#> 67    66 The Lord of the Rings: The Fellowship of the Ring          2001
#> 68    67           Fantastic Beasts and Where to Find Them          2016
#> 69    68                       Mission: Impossible—Fallout          2018
#> 70    69                                     The Lion King          1994
#> 71    70                    Jumanji: Welcome to the Jungle          2017
#> 72    71                                      Spider-Man 3          2007
#> 73    72                                      Finding Nemo          2003
#> 74    73               Star Wars Ep. I: The Phantom Menace          1999
#> 75    74                                        Mei Ren Yu          2016
#> 76    75                                Fast and Furious 6          2013
#> 77    76                                 The Da Vinci Code          2006
#> 78    77          Harry Potter and the Prisoner of Azkaban          2004
#> 79    78                            Spider-Man: Homecoming          2017
#> 80    79                                         Inception          2010
#> 81    80                                      Finding Dory          2016
#> 82    81                             Tang Ren Jie Tan An 2          2018
#> 83    82                                    Thor: Ragnarok          2017
#> 84    83          The Twilight Saga: Breaking Dawn, Part 2          2012
#> 85    84                Batman v Superman: Dawn of Justice          2016
#> 86    85                Madagascar 3: Europe's Most Wanted          2012
#> 87    86                      Rogue One: A Star Wars Story          2016
#> 88    87                           The Secret Life of Pets          2016
#> 89    88                               Shrek Forever After          2010
#> 90    89                                        Maleficent          2014
#> 91    90                        X-Men: Days of Future Past          2014
#> 92    91                                  Independence Day          1996
#> 93    92                          The Amazing Spider-Man 2          2014
#> 94    93                    Dawn of the Planet of the Apes          2014
#> 95    94                                        Inside Out          2015
#> 96    95                                   Kung Fu Panda 2          2011
#> 97    96                                        Life of Pi          2012
#> 98    97                            The Amazing Spider-Man          2012
#> 99    98                                           Shrek 2          2004
#> 100   99                  Mission: Impossible—Rogue Nation          2015
#> 101  100       Fantastic Beasts: The Crimes of Grindelwald          2018
#>     american_box_office international_box_office total_box_office
#> 2             760507625               2029197650       2789705275
#> 3             858373000               1939427564       2797800564
#> 4             659363944               1548844451       2208208395
#> 5             678815482               1369544272       2048359754
#> 6             353007020               1165715774       1518722794
#> 7             936662225               1131561399       2068223624
#> 8             543617567               1111810064       1655427631
#> 9             225764765               1009081502       1234846267
#> 10            652270625                996584239       1648854864
#> 11            381193157                960500000       1341693157
#> 12            459005868                944008095       1403013963
#> 13            623357910                895455078       1518812988
#> 14            417719760                888053039       1305772799
#> 15            400738009                871731901       1272469910
#> 16            245439076                858614996       1104054072
#> 17              2721100                855881967        858603067
#> 18            336045770                824290403       1160336173
#> 19            335061807                811832833       1146894640
#> 20            408992272                806400000       1215392272
#> 21            304360277                806166704       1110526981
#> 22            241063875                804600000       1045663875
#> 23            352390543                771403536       1123794079
#> 24            264624300                770100410       1034724710
#> 25            377845905                763557436       1141403341
#> 26            504014165                759506961       1263521126
#> 27            408084349                745200000       1153284349
#> 28            390532085                741395911       1131927996
#> 29            161321843                718443294        879765137
#> 30            303003568                714000000       1017003568
#> 31            620181382                712358507       1332539889
#> 32            328571522                709986847       1038558369
#> 33              3695533                705101664        708797197
#> 34            258366855                702000000        960366855
#> 35            426829839                701444955       1128274794
#> 36            355559216                695134737       1050693953
#> 37            341268248                695100000       1036368248
#> 38                   NA                692163684        692163684
#> 39            334191110                691300000       1025491110
#> 40            255119788                690457833        945577621
#> 41            216303339                687227217        903530556
#> 42            200074175                679546748        879620923
#> 43            296131568                664300000        960431568
#> 44            196573705                663128152        859701857
#> 45            317871467                657176139        975047606
#> 46            309420425                654000000        963420425
#> 47            415004880                653874642       1068879522
#> 48            292137260                650939197        943076457
#> 49            700059566                646853595       1346913161
#> 50            402523348                643104279       1045627627
#> 51            423315812                642900000       1066215812
#> 52            213511408                640117197        853628605
#> 53            434022296                639356585       1073378881
#> 54            448139099                636300000       1084439099
#> 55            608581744                634223615       1242805359
#> 56            302089278                633124489        935213767
#> 57            262233381                616991754        879225135
#> 58            172558876                615682261        788241137
#> 59            368065385                607151450        975216835
#> 60            290201752                606898042        897099794
#> 61            364001123                598853424        962854547
#> 62            209726015                597400000        807126015
#> 63            342548984                592150661        934699645
#> 64            166112167                591565581        757677748
#> 65            173810100                584980652        758790752
#> 66              1543547                576647204        578190751
#> 67            315544750                571666235        887210985
#> 68            234037575                568365278        802402853
#> 69            220159104                567297448        787456552
#> 70            421785283                564429585        986214868
#> 71            404508916                559987277        964496193
#> 72            336530303                558329927        894860230
#> 73            380529370                555900000        936429370
#> 74            474544677                552500000       1027044677
#> 75              3229457                551287214        554516671
#> 76            238679850                550620594        789300444
#> 77            217536138                550284321        767820459
#> 78            249757726                547149597        796907323
#> 79            334201140                545965210        880166350
#> 80            292576195                542948447        835524642
#> 81            486295561                542275328       1028570889
#> 82              1983984                542084590        544068574
#> 83            315058289                538900000        853958289
#> 84            292324737                537400000        829724737
#> 85            330360194                537140087        867500281
#> 86            216391482                530529789        746921271
#> 87            532177324                523879949       1056057273
#> 88            368384330                518222082        886606412
#> 89            238736787                517507886        756244673
#> 90            241407328                517129407        758536735
#> 91            233921534                513941241        747862775
#> 92            306169255                511231623        817400878
#> 93            202853933                506142403        708996336
#> 94            208545589                502098977        710644566
#> 95            356461711                501100000        857561711
#> 96            165249063                499588484        664837547
#> 97            124987022                495924981        620912003
#> 98            262030663                495859604        757890267
#> 99            441226247                495781885        937008132
#> 100           195042377                493816615        688858992
#> 101           159555901                492664185        652220086

top_grossing(type = "international", ranks = 1:10)

#> Please note that these numbers are not adjusted for inflation.

#>    rank                                         movie year_released
#> 2     1                                        Avatar          2009
#> 3     2                             Avengers: Endgame          2019
#> 4     3                                       Titanic          1997
#> 5     4                        Avengers: Infinity War          2018
#> 6     5                                     Furious 7          2015
#> 7     6          Star Wars Ep. VII: The Force Awakens          2015
#> 8     7                                 The Lion King          2019
#> 9     8                       The Fate of the Furious          2017
#> 10    9                                Jurassic World          2015
#> 11   10 Harry Potter and the Deathly Hallows: Part II          2011
#>    american_box_office international_box_office total_box_office
#> 2            760507625               2029197650       2789705275
#> 3            858373000               1939427564       2797800564
#> 4            659363944               1548844451       2208208395
#> 5            678815482               1369544272       2048359754
#> 6            353007020               1165715774       1518722794
#> 7            936662225               1131561399       2068223624
#> 8            543617567               1111810064       1655427631
#> 9            225764765               1009081502       1234846267
#> 10           652270625                996584239       1648854864
#> 11           381193157                960500000       1341693157