Get the top N ranking grossing movies

Get the top N ranking grossing movies

top_grossing(type = "american", ranks = 1:100)

Arguments

type
A string that says which type of box office sorted ranking you want. Options are 'american' (American box office), 'international' (non-American) and 'worldwide' (domestic + international box office).

ranks
A vector of rankings you want it to return. For example. an input of 1:5 will return the top 5 grossing movies.

Value

Data frame returning info on the name of the movie, it's rank, the year the movie was released, and the total gross from domestic (American), international, and total ticket sales.

Examples

top_grossing()

#> Please note that these numbers are not adjusted for inflation.

#>     rank                                              movie year_released
#> 2      1               Star Wars Ep. VII: The Force Awakens          2015
#> 3      2                                             Avatar          2009
#> 4      3                                      Black Panther          2018
#> 5      4                             Avengers: Infinity War          2018
#> 6      5                                            Titanic          1997
#> 7      6                                     Jurassic World          2015
#> 8      7                                       The Avengers          2012
#> 9      8                  Star Wars Ep. VIII: The Last Jedi          2017
#> 10     9                                      Incredibles 2          2018
#> 11    10                                    The Dark Knight          2008
#> 12    11                       Rogue One: A Star Wars Story          2016
#> 13    12                               Beauty and the Beast          2017
#> 14    13                                       Finding Dory          2016
#> 15    14                Star Wars Ep. I: The Phantom Menace          1999
#> 16    15                       Star Wars Ep. IV: A New Hope          1977
#> 17    16                            Avengers: Age of Ultron          2015
#> 18    17                              The Dark Knight Rises          2012
#> 19    18                                            Shrek 2          2004
#> 20    19                          ET: The Extra-Terrestrial          1982
#> 21    20                    The Hunger Games: Catching Fire          2013
#> 22    21         Pirates of the Caribbean: Dead Man’s Chest          2006
#> 23    22                                      The Lion King          1994
#> 24    23                     Jurassic World: Fallen Kingdom          2018
#> 25    24                                        Toy Story 3          2010
#> 26    25                                       Wonder Woman          2017
#> 27    26                                         Iron Man 3          2013
#> 28    27                         Captain America: Civil War          2016
#> 29    28                                   The Hunger Games          2012
#> 30    29                     Jumanji: Welcome to the Jungle          2017
#> 31    30                                         Spider-Man          2002
#> 32    31                Transformers: Revenge of the Fallen          2009
#> 33    32                                             Frozen          2013
#> 34    33                                      Jurassic Park          1993
#> 35    34                      Guardians of the Galaxy Vol 2          2017
#> 36    35      Harry Potter and the Deathly Hallows: Part II          2011
#> 37    36                                       Finding Nemo          2003
#> 38    37             Star Wars Ep. III: Revenge of the Sith          2005
#> 39    38      The Lord of the Rings: The Return of the King          2003
#> 40    39                               Beauty and the Beast          1991
#> 41    40                                       Spider-Man 2          2004
#> 42    41                          The Passion of the Christ          2004
#> 43    42                            The Secret Life of Pets          2016
#> 44    43                                    Despicable Me 2          2013
#> 45    44                                    The Jungle Book          2016
#> 46    45                                           Deadpool          2016
#> 47    46                                         Inside Out          2015
#> 48    47                                          Furious 7          2015
#> 49    48                     Transformers: Dark of the Moon          2011
#> 50    49                                    American Sniper          2014
#> 51    50              The Lord of the Rings: The Two Towers          2002
#> 52    51                                           Zootopia          2016
#> 53    52              The Hunger Games: Mockingjay - Part 1          2014
#> 54    53                                       Spider-Man 3          2007
#> 55    54                                            Minions          2015
#> 56    55                             Spider-Man: Homecoming          2017
#> 57    56                                Alice in Wonderland          2010
#> 58    57                            Guardians of the Galaxy          2014
#> 59    58                 Batman v Superman: Dawn of Justice          2016
#> 60    59                                       Forrest Gump          1994
#> 61    60                                                 It          2017
#> 62    61                                      Suicide Squad          2016
#> 63    62                                    Shrek the Third          2007
#> 64    63                                         Deadpool 2          2018
#> 65    64                                       Transformers          2007
#> 66    65                                           Iron Man          2008
#> 67    66              Harry Potter and the Sorcerer’s Stone          2001
#> 68    67 Indiana Jones and the Kingdom of the Crystal Skull          2008
#> 69    68  The Lord of the Rings: The Fellowship of the Ring          2001
#> 70    69                                     Thor: Ragnarok          2017
#> 71    70                                         Iron Man 2          2010
#> 72    71             Star Wars Ep. II: Attack of the Clones          2002
#> 73    72           Pirates of the Caribbean: At World’s End          2007
#> 74    73               Star Wars Ep. VI: Return of the Jedi          1983
#> 75    74                                   Independence Day          1996
#> 76    75   Pirates of the Caribbean: The Curse of the Blac…          2003
#> 77    76                                            Skyfall          2012
#> 78    77                  The Hobbit: An Unexpected Journey          2012
#> 79    78             Harry Potter and the Half-Blood Prince          2009
#> 80    79                         The Twilight Saga: Eclipse          2010
#> 81    80                        The Twilight Saga: New Moon          2009
#> 82    81       Harry Potter and the Deathly Hallows: Part I          2010
#> 83    82                                    The Sixth Sense          1999
#> 84    83                                                 Up          2009
#> 85    84                                          Inception          2010
#> 86    85           The Twilight Saga: Breaking Dawn, Part 2          2012
#> 87    86          Harry Potter and the Order of the Phoenix          2007
#> 88    87   The Chronicles of Narnia: The Lion, the Witch a…          2005
#> 89    88                                       Man of Steel          2013
#> 90    89           Star Wars Ep. V: The Empire Strikes Back          1980
#> 91    90                Harry Potter and the Goblet of Fire          2005
#> 92    91                                     Monsters, Inc.          2001
#> 93    92                                         Home Alone          1990
#> 94    93              The Hunger Games: Mockingjay - Part 2          2015
#> 95    94                                The Matrix Reloaded          2003
#> 96    95           The Twilight Saga: Breaking Dawn, Part 1          2011
#> 97    96                                   Meet the Fockers          2004
#> 98    97                                       The Hangover          2009
#> 99    98                                            Gravity          2013
#> 100   99                                               Sing          2016
#> 101  100                                Monsters University          2013
#>     american_box_office international_box_office total_box_office
#> 2             936662225               1116648995       2053311220
#> 3             760507625               2015837654       2776345279
#> 4             700059566                647011693       1347071259
#> 5             678815482               1370000000       2048815482
#> 6             659363944               1548844451       2208208395
#> 7             652270625                996622583       1648893208
#> 8             623279547                894656350       1517935897
#> 9             620181382                696583402       1316764784
#> 10            608581744                631980442       1240562186
#> 11            533345358                468275260       1001620618
#> 12            532177324                516998253       1049175577
#> 13            504014165                755185541       1259199706
#> 14            486295561                535055322       1021350883
#> 15            474544677                552500000       1027044677
#> 16            460998007                325600000        786598007
#> 17            459005868                944008095       1403013963
#> 18            448139099                636300000       1084439099
#> 19            441226247                495781885        937008132
#> 20            435110554                357854772        792965326
#> 21            424668047                440200000        864868047
#> 22            423315812                642900000       1066215812
#> 23            421785283                564429585        986214868
#> 24            416769345                888096977       1304866322
#> 25            415004880                653874642       1068879522
#> 26            412563408                408569970        821133378
#> 27            408992272                806400000       1215392272
#> 28            408084349                731996675       1140081024
#> 29            408010692                269912687        677923379
#> 30            404508916                560017757        964526673
#> 31            403706375                418000000        821706375
#> 32            402111870                434407829        836519699
#> 33            400738009                871731901       1272469910
#> 34            395708305                643104279       1038812584
#> 35            389813101                472504158        862317259
#> 36            381011219                960500000       1341511219
#> 37            380529370                555900000        936429370
#> 38            380270577                468728300        848998877
#> 39            377845905                763557436       1141403341
#> 40            376057266                232373866        608431132
#> 41            373524485                421586185        795110670
#> 42            370782930                251558994        622341924
#> 43            368384330                518383092        886767422
#> 44            368065385                607151450        975216835
#> 45            364001123                598871578        962872701
#> 46            363070709                437958540        801029249
#> 47            356461711                497776416        854238127
#> 48            353007020               1165715774       1518722794
#> 49            352390543                771400000       1123790543
#> 50            350126372                197200000        547326372
#> 51            342548984                592150661        934699645
#> 52            341268248                678436100       1019704348
#> 53            337135885                429516403        766652288
#> 54            336530303                558329927        894860230
#> 55            336045770                824290403       1160336173
#> 56            334201140                545869746        880070886
#> 57            334191110                691300000       1025491110
#> 58            333172112                437695404        770867516
#> 59            330360194                537143905        867504099
#> 60            330151138                349699499        679850637
#> 61            327481748                369977480        697459228
#> 62            325100054                420969188        746069242
#> 63            322719944                484610992        807330936
#> 64            322685858                415318175        738004033
#> 65            319246193                389026399        708272592
#> 66            318604126                266567421        585171547
#> 67            317575550                657176139        974751689
#> 68            317101119                469534294        786635413
#> 69            315544750                571666235        887210985
#> 70            315058289                531921735        846980024
#> 71            312433331                308723058        621156389
#> 72            310676740                346018875        656695615
#> 73            309420425                654000000        963420425
#> 74            309205079                263500000        572705079
#> 75            306169255                511231623        817400878
#> 76            305411224                329542879        634954103
#> 77            304360277                806166704       1110526981
#> 78            303003568                714000000       1017003568
#> 79            301959197                633124489        935083686
#> 80            300531751                405571077        706102828
#> 81            296623634                390934093        687557727
#> 82            295983305                664300000        960283305
#> 83            293506292                379300000        672806292
#> 84            293004164                438459213        731463377
#> 85            292576195                539825887        832402082
#> 86            292324737                537400000        829724737
#> 87            292004738                650939197        942943935
#> 88            291710957                428828615        720539572
#> 89            291045518                376954000        667999518
#> 90            290271960                243889374        534161334
#> 91            290013036                606898042        896911078
#> 92            289423425                270334294        559757719
#> 93            285761243                190923432        476684675
#> 94            281723902                367262885        648986787
#> 95            281553689                457023240        738576929
#> 96            281287133                408132918        689420051
#> 97            279167575                237400000        516567575
#> 98            277322503                188441583        465764086
#> 99            274092705                419605968        693698673
#> 100           270329045                364129526        634458571
#> 101           268488329                475100000        743588329

top_grossing(ranks = 1:5)

#> Please note that these numbers are not adjusted for inflation.

#>   rank                                movie year_released american_box_office
#> 2    1 Star Wars Ep. VII: The Force Awakens          2015           936662225
#> 3    2                               Avatar          2009           760507625
#> 4    3                        Black Panther          2018           700059566
#> 5    4               Avengers: Infinity War          2018           678815482
#> 6    5                              Titanic          1997           659363944
#>   international_box_office total_box_office
#> 2               1116648995       2053311220
#> 3               2015837654       2776345279
#> 4                647011693       1347071259
#> 5               1370000000       2048815482
#> 6               1548844451       2208208395

top_grossing(type = "international")

#> Please note that these numbers are not adjusted for inflation.

#>     rank                                              movie year_released
#> 2      1                                             Avatar          2009
#> 3      2                                            Titanic          1997
#> 4      3                             Avengers: Infinity War          2018
#> 5      4                                          Furious 7          2015
#> 6      5               Star Wars Ep. VII: The Force Awakens          2015
#> 7      6                            The Fate of the Furious          2017
#> 8      7                                     Jurassic World          2015
#> 9      8      Harry Potter and the Deathly Hallows: Part II          2011
#> 10     9                            Avengers: Age of Ultron          2015
#> 11    10                                       The Avengers          2012
#> 12    11                     Jurassic World: Fallen Kingdom          2018
#> 13    12                                             Frozen          2013
#> 14    13                    Transformers: Age of Extinction          2014
#> 15    14                                     Wolf Warrior 2          2017
#> 16    15                                            Minions          2015
#> 17    16                                         Iron Man 3          2013
#> 18    17                                            Skyfall          2012
#> 19    18        Pirates of the Caribbean: On Stranger Tides          2011
#> 20    19                     Transformers: Dark of the Moon          2011
#> 21    20                                    Despicable Me 3          2017
#> 22    21      The Lord of the Rings: The Return of the King          2003
#> 23    22                               Beauty and the Beast          2017
#> 24    23                         Captain America: Civil War          2016
#> 25    24                         Ice Age: Continental Drift          2012
#> 26    25                  The Hobbit: An Unexpected Journey          2012
#> 27    26                The Hobbit: The Desolation of Smaug          2013
#> 28    27                  Star Wars Ep. VIII: The Last Jedi          2017
#> 29    28                                Alice in Wonderland          2010
#> 30    29          The Hobbit: The Battle of the Five Armies          2014
#> 31    30                                            Spectre          2015
#> 32    31                                           Zootopia          2016
#> 33    32       Harry Potter and the Deathly Hallows: Part I          2010
#> 34    33                     Ice Age: Dawn of the Dinosaurs          2009
#> 35    34              Harry Potter and the Sorcerer’s Stone          2001
#> 36    35           Pirates of the Caribbean: At World’s End          2007
#> 37    36                                        Toy Story 3          2010
#> 38    37          Harry Potter and the Order of the Phoenix          2007
#> 39    38                                      Black Panther          2018
#> 40    39                                      Jurassic Park          1993
#> 41    40         Pirates of the Caribbean: Dead Man’s Chest          2006
#> 42    41                                              Venom          2018
#> 43    42                              The Dark Knight Rises          2012
#> 44    43             Harry Potter and the Half-Blood Prince          2009
#> 45    44                                      Incredibles 2          2018
#> 46    45                                               2012          2009
#> 47    46            Harry Potter and the Chamber of Secrets          2002
#> 48    47   Pirates of the Caribbean: Dead Men Tell No Tales          2017
#> 49    48                                    Despicable Me 2          2013
#> 50    49                Harry Potter and the Goblet of Fire          2005
#> 51    50                                    The Jungle Book          2016
#> 52    51              The Lord of the Rings: The Two Towers          2002
#> 53    52                                               Coco          2017
#> 54    53                                 Hong Hai Xing Dong          2018
#> 55    54  The Lord of the Rings: The Fellowship of the Ring          2001
#> 56    55            Fantastic Beasts and Where to Find Them          2016
#> 57    56                        Mission: Impossible—Fallout          2018
#> 58    57                                      The Lion King          1994
#> 59    58                     Jumanji: Welcome to the Jungle          2017
#> 60    59                                       Spider-Man 3          2007
#> 61    60                                       Finding Nemo          2003
#> 62    61                Star Wars Ep. I: The Phantom Menace          1999
#> 63    62                                         Mei Ren Yu          2016
#> 64    63                                 Fast and Furious 6          2013
#> 65    64                                  The Da Vinci Code          2006
#> 66    65           Harry Potter and the Prisoner of Azkaban          2004
#> 67    66                             Spider-Man: Homecoming          2017
#> 68    67                              Tang Ren Jie Tan An 2          2018
#> 69    68                                          Inception          2010
#> 70    69           The Twilight Saga: Breaking Dawn, Part 2          2012
#> 71    70                 Batman v Superman: Dawn of Justice          2016
#> 72    71                                       Finding Dory          2016
#> 73    72                                     Thor: Ragnarok          2017
#> 74    73                 Madagascar 3: Europe's Most Wanted          2012
#> 75    74                            The Secret Life of Pets          2016
#> 76    75                                Shrek Forever After          2010
#> 77    76                                         Maleficent          2014
#> 78    77                       Rogue One: A Star Wars Story          2016
#> 79    78                         X-Men: Days of Future Past          2014
#> 80    79                                   Independence Day          1996
#> 81    80                           The Amazing Spider-Man 2          2014
#> 82    81                     Dawn of the Planet of the Apes          2014
#> 83    82                                    Kung Fu Panda 2          2011
#> 84    83                                         Inside Out          2015
#> 85    84                             The Amazing Spider-Man          2012
#> 86    85                                            Shrek 2          2004
#> 87    86                   Mission: Impossible—Rogue Nation          2015
#> 88    87                 Mission: Impossible—Ghost Protocol          2011
#> 89    88                                    Shrek the Third          2007
#> 90    89                                         Life of Pi          2012
#> 91    90                                       Interstellar          2014
#> 92    91                                     Men in Black 3          2012
#> 93    92                                Monsters University          2013
#> 94    93                      Transformers: The Last Knight          2017
#> 95    94                      Guardians of the Galaxy Vol 2          2017
#> 96    95                                   Les Intouchables          2012
#> 97    96                                         Mamma Mia!          2008
#> 98    97 Indiana Jones and the Kingdom of the Crystal Skull          2008
#> 99    98             Star Wars Ep. III: Revenge of the Sith          2005
#> 100   99                                    The Dark Knight          2008
#> 101  100                                The Matrix Reloaded          2003
#>     american_box_office international_box_office total_box_office
#> 2             760507625               2015837654       2776345279
#> 3             659363944               1548844451       2208208395
#> 4             678815482               1370000000       2048815482
#> 5             353007020               1165715774       1518722794
#> 6             936662225               1116648995       2053311220
#> 7             225764765               1009143255       1234908020
#> 8             652270625                996622583       1648893208
#> 9             381011219                960500000       1341511219
#> 10            459005868                944008095       1403013963
#> 11            623279547                894656350       1517935897
#> 12            416769345                888096977       1304866322
#> 13            400738009                871731901       1272469910
#> 14            245439076                858600000       1104039076
#> 15              2721100                855881967        858603067
#> 16            336045770                824290403       1160336173
#> 17            408992272                806400000       1215392272
#> 18            304360277                806166704       1110526981
#> 19            241063875                804600000       1045663875
#> 20            352390543                771400000       1123790543
#> 21            264624300                770123890       1034748190
#> 22            377845905                763557436       1141403341
#> 23            504014165                755185541       1259199706
#> 24            408084349                731996675       1140081024
#> 25            161321843                718443294        879765137
#> 26            303003568                714000000       1017003568
#> 27            258366855                702000000        960366855
#> 28            620181382                696583402       1316764784
#> 29            334191110                691300000       1025491110
#> 30            255119788                690750000        945869788
#> 31            200074175                679546748        879620923
#> 32            341268248                678436100       1019704348
#> 33            295983305                664300000        960283305
#> 34            196573705                663128152        859701857
#> 35            317575550                657176139        974751689
#> 36            309420425                654000000        963420425
#> 37            415004880                653874642       1068879522
#> 38            292004738                650939197        942943935
#> 39            700059566                647011693       1347071259
#> 40            395708305                643104279       1038812584
#> 41            423315812                642900000       1066215812
#> 42            212969264                640543202        853512466
#> 43            448139099                636300000       1084439099
#> 44            301959197                633124489        935083686
#> 45            608581744                631980442       1240562186
#> 46            166112167                622296372        788408539
#> 47            261987880                616991754        878979634
#> 48            172558876                615682261        788241137
#> 49            368065385                607151450        975216835
#> 50            290013036                606898042        896911078
#> 51            364001123                598871578        962872701
#> 52            342548984                592150661        934699645
#> 53            209726015                589738726        799464741
#> 54              1543547                576647204        578190751
#> 55            315544750                571666235        887210985
#> 56            234037575                568400093        802437668
#> 57            220159104                567338413        787497517
#> 58            421785283                564429585        986214868
#> 59            404508916                560017757        964526673
#> 60            336530303                558329927        894860230
#> 61            380529370                555900000        936429370
#> 62            474544677                552500000       1027044677
#> 63              3229457                551287214        554516671
#> 64            238679850                550620594        789300444
#> 65            217536138                550284321        767820459
#> 66            249538952                547149597        796688549
#> 67            334201140                545869746        880070886
#> 68              1983984                542084590        544068574
#> 69            292576195                539825887        832402082
#> 70            292324737                537400000        829724737
#> 71            330360194                537143905        867504099
#> 72            486295561                535055322       1021350883
#> 73            315058289                531921735        846980024
#> 74            216391482                530529789        746921271
#> 75            368384330                518383092        886767422
#> 76            238736787                517507886        756244673
#> 77            241407328                517129407        758536735
#> 78            532177324                516998253       1049175577
#> 79            233921534                513941241        747862775
#> 80            306169255                511231623        817400878
#> 81            202853933                506142403        708996336
#> 82            208545589                502098977        710644566
#> 83            165249063                499588484        664837547
#> 84            356461711                497776416        854238127
#> 85            262030663                495859604        757890267
#> 86            441226247                495781885        937008132
#> 87            195042377                493822021        688864398
#> 88            209397903                485315327        694713230
#> 89            322719944                484610992        807330936
#> 90            124987022                482271612        607258634
#> 91            188017894                479734528        667752422
#> 92            179020854                475192631        654213485
#> 93            268488329                475100000        743588329
#> 94            130168683                472724657        602893340
#> 95            389813101                472504158        862317259
#> 96             13182281                471690764        484873045
#> 97            144130063                471618709        615748772
#> 98            317101119                469534294        786635413
#> 99            380270577                468728300        848998877
#> 100           533345358                468275260       1001620618
#> 101           281553689                457023240        738576929

top_grossing(type = "international", ranks = 1:10)

#> Please note that these numbers are not adjusted for inflation.

#>    rank                                         movie year_released
#> 2     1                                        Avatar          2009
#> 3     2                                       Titanic          1997
#> 4     3                        Avengers: Infinity War          2018
#> 5     4                                     Furious 7          2015
#> 6     5          Star Wars Ep. VII: The Force Awakens          2015
#> 7     6                       The Fate of the Furious          2017
#> 8     7                                Jurassic World          2015
#> 9     8 Harry Potter and the Deathly Hallows: Part II          2011
#> 10    9                       Avengers: Age of Ultron          2015
#> 11   10                                  The Avengers          2012
#>    american_box_office international_box_office total_box_office
#> 2            760507625               2015837654       2776345279
#> 3            659363944               1548844451       2208208395
#> 4            678815482               1370000000       2048815482
#> 5            353007020               1165715774       1518722794
#> 6            936662225               1116648995       2053311220
#> 7            225764765               1009143255       1234908020
#> 8            652270625                996622583       1648893208
#> 9            381011219                960500000       1341511219
#> 10           459005868                944008095       1403013963
#> 11           623279547                894656350       1517935897